Remove initial poc for 3 step checkout and rename beta to standard as it's rolled out to everyone now.
