--- src/pages/Checkout/BasketSummary.vue
@@ -1,11 +1,0 @@
-<template>
-	<div class="basket">
-		<h1>Basket</h1>
-	</div>
-</template>
-
-<script>
-export default {
-
-};
-</script>
