--- src/pages/Checkout/CheckoutItemsList.vue
@@ -1,17 +1,0 @@
-<template>
-	<div>
-		<ul>
-			<li class=""></li>
-			<li class=""></li>
-			<li class=""></li>
-		</ul>
-	</div>
-</template>
-
-<script>
-
-</script>
-
-<style lang="scss">
-@import 'settings';
-</style>
