--- src/pages/Checkout/PaymentSummary.vue
@@ -1,11 +1,0 @@
-<template>
-	<div class="payment">
-		<h1>Payment</h1>
-	</div>
-</template>
-
-<script>
-export default {
-
-};
-</script>
