--- src/pages/Checkout/ThanksSummary.vue
@@ -1,11 +1,0 @@
-<template>
-	<div class="thanks">
-		<h1>Thanks</h1>
-	</div>
-</template>
-
-<script>
-export default {
-
-};
-</script>
