+++ src/router/routes.js
@@ -9,8 +9,7 @@ export default [
 	{ path: '/page-two', component: () => import('@/pages/PageTwo') },
 	{ path: '/UI_REVISION', component: () => import('@/pages/UiRevision') },
 	{ path: '/login-vue', component: () => import('@/pages/LoginAndRegister/LoginAndRegister') },
-	{ path: '/checkout-vue', component: () => import('@/pages/Checkout/CheckoutPage') },
-	{ path: '/checkout', component: () => import('@/pages/Checkout/CheckoutBetaPage') },
+	{ path: '/checkout', component: () => import('@/pages/Checkout/CheckoutPage') },
 	{ path: '/protocol', component: () => import('@/pages/Protocol') },
 	{ path: '/build', component: () => import('@/pages/Build/BuildPage') },
 	{ path: '/build/docs', component: () => import('@/pages/Build/DocsPage') },
